Expose the client utils for external integrations
