+++ pkg/client/initrequest_lister.go
@@ -14,21 +14,19 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package initrequest
+package client
 
 import (
 	"context"
 	"fmt"
 
 	directpvtypes "github.com/minio/directpv/pkg/apis/directpv.min.io/types"
-	"github.com/minio/directpv/pkg/client"
 	"github.com/minio/directpv/pkg/k8s"
 	"github.com/minio/directpv/pkg/types"
 	"github.com/minio/directpv/pkg/utils"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/watch"
 )
 
 // ListInitRequestResult denotes list of initrequest result.
@@ -37,8 +35,8 @@ type ListInitRequestResult struct {
 	Err         error
 }
 
-// Lister is initRequest lister.
-type Lister struct {
+// InitRequestLister is initRequest lister.
+type InitRequestLister struct {
 	nodes            []directpvtypes.LabelValue
 	requestIDs       []directpvtypes.LabelValue
 	initRequestNames []string
@@ -46,45 +44,45 @@ type Lister struct {
 	ignoreNotFound   bool
 }
 
-// NewLister creates new volume lister.
-func NewLister() *Lister {
-	return &Lister{
+// NewInitRequestLister creates new volume lister.
+func NewInitRequestLister() *InitRequestLister {
+	return &InitRequestLister{
 		maxObjects: k8s.MaxThreadCount,
 	}
 }
 
 // NodeSelector adds filter listing by nodes.
-func (lister *Lister) NodeSelector(nodes []directpvtypes.LabelValue) *Lister {
+func (lister *InitRequestLister) NodeSelector(nodes []directpvtypes.LabelValue) *InitRequestLister {
 	lister.nodes = nodes
 	return lister
 }
 
 // RequestIDSelector adds filter listing by its request IDs.
-func (lister *Lister) RequestIDSelector(requestIDs []directpvtypes.LabelValue) *Lister {
+func (lister *InitRequestLister) RequestIDSelector(requestIDs []directpvtypes.LabelValue) *InitRequestLister {
 	lister.requestIDs = requestIDs
 	return lister
 }
 
 // InitRequestNameSelector adds filter listing by InitRequestNames.
-func (lister *Lister) InitRequestNameSelector(initRequestNames []string) *Lister {
+func (lister *InitRequestLister) InitRequestNameSelector(initRequestNames []string) *InitRequestLister {
 	lister.initRequestNames = initRequestNames
 	return lister
 }
 
 // MaxObjects controls number of items to be fetched in every iteration.
-func (lister *Lister) MaxObjects(n int64) *Lister {
+func (lister *InitRequestLister) MaxObjects(n int64) *InitRequestLister {
 	lister.maxObjects = n
 	return lister
 }
 
 // IgnoreNotFound controls listing to ignore not found error.
-func (lister *Lister) IgnoreNotFound(b bool) *Lister {
+func (lister *InitRequestLister) IgnoreNotFound(b bool) *InitRequestLister {
 	lister.ignoreNotFound = b
 	return lister
 }
 
 // List returns channel to loop through initrequest items.
-func (lister *Lister) List(ctx context.Context) <-chan ListInitRequestResult {
+func (lister *InitRequestLister) List(ctx context.Context) <-chan ListInitRequestResult {
 	getOnly := len(lister.nodes) == 0 &&
 		len(lister.requestIDs) == 0 &&
 		len(lister.initRequestNames) != 0
@@ -114,7 +112,7 @@ func (lister *Lister) List(ctx context.Context) <-chan ListInitRequestResult {
 				LabelSelector: labelSelector,
 			}
 			for {
-				result, err := client.InitRequestClient().List(ctx, options)
+				result, err := InitRequestClient().List(ctx, options)
 				if err != nil {
 					send(ListInitRequestResult{Err: err})
 					return
@@ -148,7 +146,7 @@ func (lister *Lister) List(ctx context.Context) <-chan ListInitRequestResult {
 		}
 
 		for _, initRequestName := range lister.initRequestNames {
-			initRequest, err := client.InitRequestClient().Get(ctx, initRequestName, metav1.GetOptions{})
+			initRequest, err := InitRequestClient().Get(ctx, initRequestName, metav1.GetOptions{})
 			if err != nil {
 				send(ListInitRequestResult{Err: err})
 				return
@@ -163,7 +161,7 @@ func (lister *Lister) List(ctx context.Context) <-chan ListInitRequestResult {
 }
 
 // Get returns list of initrequest.
-func (lister *Lister) Get(ctx context.Context) ([]types.InitRequest, error) {
+func (lister *InitRequestLister) Get(ctx context.Context) ([]types.InitRequest, error) {
 	ctx, cancelFunc := context.WithCancel(ctx)
 	defer cancelFunc()
 
@@ -178,20 +176,13 @@ func (lister *Lister) Get(ctx context.Context) ([]types.InitRequest, error) {
 	return initRequestList, nil
 }
 
-// WatchEvent represents the initrequest events.
-type WatchEvent struct {
-	Type        watch.EventType
-	InitRequest *types.InitRequest
-	Err         error
-}
-
 // Watch looks for changes in InitRequestList and reports them.
-func (lister *Lister) Watch(ctx context.Context) (<-chan WatchEvent, func(), error) {
+func (lister *InitRequestLister) Watch(ctx context.Context) (<-chan WatchEvent[*types.InitRequest], func(), error) {
 	labelMap := map[directpvtypes.LabelKey][]directpvtypes.LabelValue{
 		directpvtypes.NodeLabelKey:      lister.nodes,
 		directpvtypes.RequestIDLabelKey: lister.requestIDs,
 	}
-	initRequestWatchInterface, err := client.InitRequestClient().Watch(ctx, metav1.ListOptions{
+	initRequestWatchInterface, err := InitRequestClient().Watch(ctx, metav1.ListOptions{
 		LabelSelector: directpvtypes.ToLabelSelector(labelMap),
 	})
 	if err != nil {
@@ -199,7 +190,7 @@ func (lister *Lister) Watch(ctx context.Context) (<-chan WatchEvent, func(), err
 	}
 	stopFn := initRequestWatchInterface.Stop
 
-	watchCh := make(chan WatchEvent)
+	watchCh := make(chan WatchEvent[*types.InitRequest])
 	go func() {
 		defer close(watchCh)
 		resultCh := initRequestWatchInterface.ResultChan()
@@ -212,7 +203,7 @@ func (lister *Lister) Watch(ctx context.Context) (<-chan WatchEvent, func(), err
 			var initRequest types.InitRequest
 			err := runtime.DefaultUnstructuredConverter.FromUnstructured(unstructured.Object, &initRequest)
 			if err != nil {
-				watchCh <- WatchEvent{
+				watchCh <- WatchEvent[*types.InitRequest]{
 					Type: result.Type,
 					Err:  fmt.Errorf("unable to convert unstructured object %s; %v", unstructured.GetName(), err),
 				}
@@ -221,9 +212,9 @@ func (lister *Lister) Watch(ctx context.Context) (<-chan WatchEvent, func(), err
 			if len(lister.initRequestNames) > 0 && !utils.Contains(lister.initRequestNames, initRequest.Name) {
 				continue
 			}
-			watchCh <- WatchEvent{
-				Type:        result.Type,
-				InitRequest: &initRequest,
+			watchCh <- WatchEvent[*types.InitRequest]{
+				Type: result.Type,
+				Item: &initRequest,
 			}
 		}
 	}()
